requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
inflection==0.5.1
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
